refactor: moved tree view helpers function into separate files
